Fixed openmv riscv issues, include 8 bytes alignment and volatile float

+++ src/openmv/src/omv/img/collections.c
@@ -4,6 +4,7 @@
  */
 
 #include "imlib.h"
+
 #define CHAR_BITS (sizeof(char) * 8)
 #define CHAR_MASK (CHAR_BITS - 1)
 #define CHAR_SHIFT IM_LOG2(CHAR_MASK)
@@ -54,7 +55,7 @@ void lifo_alloc(lifo_t *ptr, size_t size, size_t data_len)
 
 void lifo_alloc_all(lifo_t *ptr, size_t *size, size_t data_len)
 {
-    uint32_t tmp_size;
+    uint64_t tmp_size;
     ptr->data = (char *) fb_alloc_all(&tmp_size);
     ptr->data_len = data_len;
     ptr->size = tmp_size / data_len;
@@ -131,7 +132,7 @@ void fifo_alloc(fifo_t *ptr, size_t size, size_t data_len)
 
 void fifo_alloc_all(fifo_t *ptr, size_t *size, size_t data_len)
 {
-    uint32_t tmp_size;
+    uint64_t tmp_size;
     ptr->data = (char *) fb_alloc_all(&tmp_size);
     ptr->data_len = data_len;
     ptr->size = tmp_size / data_len;
